requirements.txt
# Snake RL Project - Windows CPU Development Requirements
# Install: pip install -r requirements.txt

# Core RL & ML
gymnasium>=0.29.0,<0.30.0
stable-baselines3[extra]>=2.3.0,<2.4.0
# SB3 extras includes: tensorboard, opencv-python, pandas, matplotlib

# PyTorch CPU (Windows-compatible)
--extra-index-url https://download.pytorch.org/whl/cpu
torch>=2.2.0,<3.0.0
torchvision>=0.17.0,<1.0.0

# Core dependencies
numpy>=1.26.0,<2.0.0
matplotlib>=3.8.0
tensorboard>=2.17.0
pandas>=2.0.0
pyyaml>=6.0.0
tqdm>=4.66.0

# Visualization & monitoring
opencv-python>=4.9.0
psutil>=5.9.0
pillow>=10.0.0

# Dashboard
streamlit>=1.32.0
gradio>=4.20.0

# Code quality
pytest>=8.0.0
pytest-cov>=4.1.0
black>=24.0.0
ruff>=0.3.0
mypy>=1.9.0
pre-commit>=3.6.0

# Jupyter for analysis
jupyter>=1.0.0
ipykernel>=6.29.0
seaborn>=0.13.0
